getting character plus pages works now so that you dont make an API request for every page but instead it saves which character is on which page

+++ src/main/kotlin/net/grandcentrix/backend/models/ResponseData.kt
@@ -46,6 +46,7 @@ data class MovieDataItem(
 )
 
 
+
 @Serializable
 data class BookResponseData(
     val data: BookDataItem
@@ -93,8 +94,8 @@ data class PaginationData(
     val first: Int? = null,
     val previous: Int? = null,
     val next: Int? = null,
-    val last: Int? = null,
-    val records: Int? = null
+    val last: Long? = null,
+    val records: Long? = null
 )
 
 @Serializable
